Add repository, service, and controller classes

+++ autoskola/src/main/java/com/example/autoskola/repository/CandidateRepository.java
@@ -3,7 +3,7 @@ package com.example.autoskola.repository;
 import com.example.autoskola.model.Candidate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CandidateRepository extends JpaRepository<Candidate, Integer> {
+public interface CandidateRepository extends JpaRepository<Candidate, Long> {
 
         Candidate save(Candidate candidate);
 }
